# Dodatne datoteke/openapi.yaml
openapi: 3.0.3
info:
  title: FER dvorane
  description: API za FER dvorane bazu podataka
  license:
    name: Creative Commons Zero v1.0 Universal
    url: https://creativecommons.org/publicdomain/zero/1.0/
  version: 1.0.0
paths:
  /zgrade:
    get:
      tags:
        - Zgrade
      summary: Popis zgrada s atributima
      description: Vraća popis zgrada s atributima
      responses:
        "200":
          description: Successful response
    post:
      tags:
        - Zgrade
      summary: Dodavanje zgrade
      description: Dodaje zgradu iz tijela zahtjeva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oznzgrada:
                  type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  oznZgrada:
                    type: string
                  dvorane:
                    type: array
                    items:
                      type: object
        "400":
          description: Bad request
        "409":
          description: Confilct - that resource already exists
  /zgrade/{oznzgrada}:
    get:
      tags:
        - Zgrade
      summary: Zgrada s atributima
      description: Vraća zgradu s atributima temeljem oznake
      parameters:
        - in: path
          name: oznzgrada
          schema:
            type: string
          required: true
          description: Oznaka zgrade koju treba dohvatiti
      responses:
        "200":
          description: Successful response
        "404":
          description: Not Found - invalid oznzgrada
    put:
      tags:
        - Zgrade
      summary: Ažuriranje zgrade
      description: Ažuriranje zgrade temeljem oznake i tijela zahtjeva
      parameters:
        - in: path
          name: oznzgrada
          schema:
            type: string
          required: true
          description: Oznaka zgrade koju treba ažurirati
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oznzgrada:
                  type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  oznZgrada:
                    type: string
                  dvorane:
                    type: array
                    items:
                      type: object
        "400":
          description: Bad request
        "404":
          description: Not Found - invalid oznzgrada
    delete:
      tags:
        - Zgrade
      summary: Brisanje zgrade
      description: Briše zgradu temeljem oznake
      parameters:
        - in: path
          name: oznzgrada
          schema:
            type: string
          required: true
          description: Oznaka zgrade koju treba obrisati
      responses:
        "200":
          description: Successful response
        "404":
          description: Not Found - invalid oznzgrada
  /dvorane:
    get:
      tags:
        - Dvorane
      summary: Popis dvorana s atributima
      description: Vraća popis dvorana s atributima
      parameters:
        - in: query
          required: false
          name: imaklimu
          description: Vrijednost ima li dvorana klimu
          schema:
            type: boolean
      responses:
        "200":
          description: Successful response
  /dvorane/{ozndvorana}:
    get:
      tags:
        - Dvorane
      summary: Dvorana s atributima
      description: Vraća dvoranu s atributima
      parameters:
        - in: path
          name: ozndvorana
          required: true
          description: Oznaka dvorane koju treba dohvatiti
          schema:
            type: string
      responses:
        "200":
          description: Successful response
        "404":
          description: Not Found - invalid ozndvorana